Split proxy configuration out of resource handlers
