+++ pkg/grizzly/embed/assets/favicon.ico
