+++ pkg/grizzly/embed/templates/proxy/index.html.tmpl
@@ -4,6 +4,7 @@
     <meta charset=utf-8>
     <title>Grizzly</title>
     <link rel="stylesheet" href="/grizzly/assets/style.css"/>
+    <link rel="icon" href="/grizzly/assets/favicon.ico">
 </head>
 <body dir="ltr">
 {{ template "proxy/header.html.tmpl" . }}
