+++ pkg/grizzly/utils.go
@@ -2,8 +2,6 @@ package grizzly
 
 import (
 	"fmt"
-	"log"
-	"net/http"
 )
 
 // Pluraliser returns a string describing the count of items, with a plural 's'
@@ -14,9 +12,4 @@ func Pluraliser(count int, name string) string {
 	}
 
 	return fmt.Sprintf("%d %ss", count, name)
-}
-
-func SendError(w http.ResponseWriter, msg string, err error, code int) {
-	http.Error(w, msg, code)
-	log.Printf("%d - %s: %s", code, msg, err.Error())
 }
